Added: Support for ByteBuffer source; use int length.

+++ src/VCDiff/Decoders/BodyDecoder.cs
@@ -317,7 +317,7 @@ namespace VCDiff.Decoders
             // Copy all data from source segment
             if (decodedAddress + size <= window.SourceSegmentLength)
             {
-                source.Position = decodedAddress + window.SourceSegmentOffset;
+                source.Position = (int)(decodedAddress + window.SourceSegmentOffset);
                 targetData.Write(source.ReadBytesAsSpan(size));
                 this.TotalBytesDecoded += size;
                 return VCDiffResult.SUCCESS;
@@ -328,7 +328,7 @@ namespace VCDiff.Decoders
             {
                 // ... plus some data from source segment
                 long partialCopySize = window.SourceSegmentLength - decodedAddress;
-                source.Position = decodedAddress + +window.SourceSegmentOffset;
+                source.Position = (int)(decodedAddress + +window.SourceSegmentOffset);
                 targetData.Write(source.ReadBytesAsSpan((int)partialCopySize));
                 this.TotalBytesDecoded += partialCopySize;
                 decodedAddress += partialCopySize;
